chore(live): rename dir streams to live
