+++ pages/live/[id].tsx
